pyproject.toml
# Poetry setup: for local development amd NaturalAntibody projects using riot_na as dependency

[tool.poetry]
authors = ["NaturalAntibody"]
description = ""
include = [{ path = "riot_na/riot_na.*.so", format = "wheel" }]
name = "riot_na"
version = "0.0.0"

[tool.poetry.dependencies]
blosum = "^2.0.2"
click = "^8.1.7"
maturin = "^1.3.2"
psutil = "^5.9.6"
python = "^3.9" # Only for NA projects compatibility reasons, minimum required python is 3.10 !
python-dotenv = "^1.0.0"
scikit-bio = "^0.5.9"
scipy = "^1.9"
tqdm = "^4.66.1"
biopython = "1.79"

[tool.poetry.group.dev.dependencies]
black = "^23.12.0"
commitizen = "^3.12.0"
cutadapt = "^4.5"
flake8 = "^6.1.0"
fsspec = "^2023.9.2"
ipykernel = "^6.25.2"
isort = "^5.12.0"
mypy = "^1.7.1"
pre-commit = "^3.5.0"
presto = "^0.7.1"
pyarrow = "^13.0.0"
pylint = "^3.0.2"
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
pytest-icdiff = "^0.9"
python-semantic-release = "^9.4.0"
s3fs = "^2023.9.2"
types-psutil = "^5.9.5.17"
types-tqdm = "^4.66.0.5"
tabulate = "^0.9.0"

[tool.poetry.scripts]
riot_na = "riot_na.cli:run_riot"

[build-system]
build-backend = "maturin"
requires = ["maturin>=1,<2"]

# Maturin setup: for wheel building

[project]
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
]
dependencies = [
    "biopython ~= 1.81",
    "blosum ~= 2.0.2",
    "click ~=8.1.7",
    "psutil ~=5.9.6",
    "python-dotenv ~= 1.0.0",
    "scikit-bio ~= 0.5.9",
    "scipy ~=1.9",
    "tqdm ~= 4.66.1",
]
name = "riot_na"
requires-python = ">=3.10"
version = "0.0.0"

[project.scripts]
riot_na = "riot_na.cli:run_riot"

[tool.maturin]
include = [{ path = "riot", format = "wheel" }]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.semantic_release]
build_command = "maturin build -r --out dist/"
changelog_file = "CHANGELOG.md"
commit_message = "bump: {version} [skip ci]\n\nAutomatically generated by python-semantic-release"
commit_version_number = true
hvcs = "github"
tag_format = "v{version}"
upload_to_pypi = false
version_toml = [
    "pyproject.toml:tool.poetry.version",
    "pyproject.toml:project.version",
]
version_variables = ["riot_na/__init__.py:__version__"]
